added ToDefReturnType to TypeConvertLibrary to unify the return type on python's def line for both property and method.
